Refactor gather step into its own module.
This is the first of 3 refactors. Essentially the 3 classic steps of
input, process, and output are being used, though I am calling them
gather, process, and render, as we are gathering the css files,
extracting the comments and then rendering the documentation.

Aside from making the code a little easier to get around, putting the
gather methods into their own module means we can easily unit
test them.